// backend/src/dtos/auth.dto.ts
import { z } from "zod";

export const registerDto = z.object({
  name: z.string().min(2),
  email: z.string().email(),
  password: z.string().min(6),
});

export const loginDto = z.object({
  email: z.string().email(),
  password: z.string().min(6),
});

export const updateProfileDto = z.object({
  name: z.string().min(2),
});